Refactor top-level macro code
* explicitly check that we support the rhs form before doing anything
  else
* read the expected file contents explicitly.  Don't create it without
  asking user if it's not yet present
* standardize naming of variables and parameters
* more correctly handle stripping extra newline we add at the end of
  expected to make it pretty in the heredoc

Create a helper AssertValue.ask_user_about_diff to share code between
create and update case.  Whether it will make sense to keep this helper
down the line or inline code in the caller remains to be seen.  If we do
keep it, the way we pass arguments into it as opts[] is probably
temporary -- a convenient way to handle it flexibly while the format is
being standardized.