Move integration tests that answer "n" to the end.
This includes one test that should normally pass without asking
anything.  The issue was that when it broke, e.g. because we broke
assert_value, the tests would hang because we now have an extra prompt
we didn't expect.

With the new setup we just pass an extra "n" to Porcelain in normal
case, which is harmless.  And if this assert_value ends up prompting for
some reason, the tests diff but don't hang.