Better handling for strings without trailing newline
To work as a heredoc a string must end with a newline.  For strings that
don't we append a special token and a newline when writing them to
source file.  This way we can look for this special token when we read
it back and strip it at that time.

Also, better handle trailing newline when splitting strings.  Elixir
doesn't trim empty strings on split by default, so this code worked
accidentally before.