Better parse expected
Recursively add one character in loop starting from the calculated
beginning of the expected value until we match expected AST got from
macro. As a result we get expected value as it is formatted in test
code and can replace it.
Now assert_value accepts both strings and heredocs as right argument.
On update string argument is formatted as heredoc.