Better check for not-serializable values
Format actual with our formatter and evaluate formatted result.
If evaluated value is the same as it was before formatting then
it is serializable and formatted correctly.
Now we can check for not-serializable values nested in another types
like List, Map, etc.
This also makes extra check for "#" in Formatter unnecessary.