Replace the group_leader with our capturing StringIO.
By setting the group_leader to our StringIO PID, we can decide exactly
when the output is flushed to the screen while we interact with the user.

We must ensure that the formatted error from a failed assert_value
in one test is not printed after we request a diff in the next test,
otherwise their outputs get mixed together.

We tried this approach before but abandoned it because a simple
Process.sleep was enough. Now that we’ve removed it, we need
this solution again.

This time, we restore the previous group_leader when the process
terminates so we don’t lose any output at the end of the tests.